// apps/docs/types/types.ts
export type ColumnDef<T> = {
  filter?: FilterType<T>;
  cell?: (info: CellContext<T, unknown>) => any;
  sortField?: string;
  className?: string;
  headerClassName?: string;
  footerClassName?: string;
  reorderable?: boolean;

  // Column sizing properties
  size?: number;
  minSize?: number;
  maxSize?: number;
  enableResizing?: boolean;
} & (
  | (
      | {
          accessorFn: (row: T) => any;
          id: string;
          accessorKey?: never;
          columns?: never;
        }
      | {
          accessorKey: DeepKeys<T>;
          accessorFn?: never;
          id?: string;
          columns?: never;
        }
    )
  | {
      columns: ColumnDef<T>[];
      accessorFn?: never;
      accessorKey?: never;
      id?: string;
    }
) &
  TSColumnDef<T, unknown>;

export type Column<T> = {
  columnDef: ColumnDef<T>;
} & TSColumn<T, unknown>;

export type GroupColumnDef<T> = {
  columns?: ColumnDef<T>[];
} & TSGroupColumnDef<T, unknown>;

export type FilterType<T> = {
  field: string;
  className?: string;
} & (
  | {
      type: "text";
      placeholder: string;
      showList?: boolean;
      showTotal?: boolean;
    }
  | {
      type: "range";
      minPlaceholder?: string;
      maxPlaceholder?: string;
      showLimit?: boolean;
      minLimit?: number | "faceted";
      maxLimit?: number | "faceted";
    }
  | {
      type: "select" | "multi-select";
      options: any[];
      optionLabel: string;
      optionValue: string;
      allLabel?: string;
    }
  | {
      type: "boolean";
      trueLabel?: string;
      falseLabel?: string;
      allLabel?: string;
    }
  | {
      type: "date";
    }
  | {
      type: "date-range";
    }
  | {
      type: "custom";
      component: React.ComponentType<{ column: Column<T> }>;
    }
);